vertical movement down
added swipe touch onDown method  event that checks to see if any
movement of the users input is greater than x pixels. if true it calls
a restartShip method which stops the current vertical tween and
replaces it with a downward tween and sends the ships position back to
starting position.

there are 4 things that happen when we restart the ship:
1. we set canSwipe to false because the ship is already reacting to a
swipe.
2.then we stop the vertical tween
3. then we tween quickly down
4. then once the ship is completely down we restart the tween back up.