vertical movement down (more comments)
more comments to vertical down movement commit. the restartShip method
after tweeting the ship’s verticalTween property back to the original
position of 860 will use the onComplete method to tween the ship back
to the verticalTween value of y:0. (reseting the ship and setting
canSwipe back to false. which completes the loop and enables for
another swipe event.)