Barrier
Applied the Barrier Class and Group to the game.  Learned the Call
method to inherit from the Phaser.Sprite Class and used it to add
Barrier(this) as a sprite image.  Also learned the crop method which
takes a rectangle object as a parameter then crops a sprite to the
rectangles size