Customizing
changed the gameplay by letting the ship idle on Y position instead of
having the ship move to the top of the screen, i found it annoying to
switch back and forth without getting hit by barriers from behind.
Also added the game over state with a button that resets the game and
some text that will say something relevant to the style of game I
choose.  For now i added a car sprite to replace the boring arrow
sprite, that may change in the future.